chore: address Copilot review
