+++ JavaScriptRuntime/CommonJS/ModuleName.cs
@@ -67,6 +67,47 @@ namespace JavaScriptRuntime.CommonJS
             return SanitizeModuleId(relative);
         }
 
+        public static string GetModuleIdForManifestFromPath(string modulePath, string rootModulePath)
+        {
+            if (modulePath == null)
+            {
+                throw new ArgumentNullException(nameof(modulePath));
+            }
+
+            if (rootModulePath == null)
+            {
+                throw new ArgumentNullException(nameof(rootModulePath));
+            }
+
+            // Mirror GetModuleIdFromPath but stop BEFORE sanitization.
+            // This preserves path-like module ids for host-facing discovery (e.g. "calculator/index").
+            var rootFullPath = Path.GetFullPath(rootModulePath);
+            var rootDirectory = Path.GetDirectoryName(rootFullPath) ?? ".";
+
+            var moduleFullPath = Path.GetFullPath(modulePath);
+            var relative = Path.GetRelativePath(rootDirectory, moduleFullPath);
+            relative = relative.Replace('\\', '/');
+
+            if (relative.EndsWith(".js", StringComparison.OrdinalIgnoreCase))
+            {
+                relative = relative.Substring(0, relative.Length - 3);
+            }
+            else
+            {
+                // Path.ChangeExtension handles removing extension, but keep forward slashes.
+                relative = Path.ChangeExtension(relative.Replace('/', Path.DirectorySeparatorChar), null) ?? relative;
+                relative = relative.Replace('\\', '/');
+            }
+
+            if (string.IsNullOrWhiteSpace(relative))
+            {
+                throw new InvalidOperationException(
+                    $"Computed an invalid module ID from path '{modulePath}' relative to root module '{rootModulePath}'.");
+            }
+
+            return relative;
+        }
+
         private static string SanitizeModuleId(string moduleId)
         {
             if (string.IsNullOrWhiteSpace(moduleId))
